move tests and optional dependencies into the utils
