+++ src/cogsworth/pop.py
@@ -31,9 +31,8 @@ from cogsworth.kicks import integrate_orbit_with_events
 from cogsworth.events import identify_events
 from cogsworth.classify import determine_final_classes
 from cogsworth.obs.observables import get_photometry
-from cogsworth.tests.optional_deps import check_dependencies
 from cogsworth.plot import plot_cartoon_evolution, plot_galactic_orbit, plot_hrd
-from cogsworth.utils import translate_COSMIC_tables, get_default_BSE_settings
+from cogsworth.utils import translate_COSMIC_tables, get_default_BSE_settings, check_dependencies
 
 from cogsworth.citations import CITATIONS
